"from globalminifier to minifyall and to the index.ts file"
