+++ src/index.ts
@@ -1,47 +1,102 @@
-import {GlobalMinifierClass} from "./controller/globalMinifier";
-import {displayException} from "./cli/displayException";
+import {CssMinifier} from "./controller/langDefaultMinifiers/cssMinifier";
+import {JsonMinifier} from "./controller/langDefaultMinifiers/jsonMinifier";
+import {HtmlMinifier} from "./controller/langDefaultMinifiers/htmlMinifier";
+import {HexMinifier} from "./controller/hexMinifier";
+import {CommentRemover} from "./controller/commentRemover";
 
 /**
- * Function that detects the language given by the path and returns the minimized code.
- *
- * @param path the path to look for the file to be minified.
- * @param content the non minified code.
- * @param willMinifyHex if the user will minify hex and rgb values.
+ * MinifyAllClass contain functions that are used by multiple languages to be minified.
  */
-export function detectLanguageAndMinify(path: string, content: string[], willMinifyHex: boolean): string {
-    const minifier: GlobalMinifierClass = new GlobalMinifierClass(willMinifyHex);
-    switch (path.substr(path.lastIndexOf('.'), path.length)) {
-        case LanguagesEnum.HTML:
-            return minifier.minifyHtml(content);
-        case LanguagesEnum.CSS:
-            return minifier.minifyCssScssLessSass(content);
-        case LanguagesEnum.JSON:
-        case LanguagesEnum.JSONC:
-            return minifier.minifyJsonJsonc(content);
-        default:
-            displayException(402, 'file extension not recognized', path);
-            break;
+export class MinifyAllClass {
+
+    /**
+     * Constructor of the MinifyAllClass that receives the willMinifyHex boolean.
+     *
+     * @param willMinifyHex if the minifier will minify hexadecimal or RGB/A values.
+     */
+    constructor(private willMinifyHex: boolean) {
     }
-}
 
-/**
- * Enum of the languages available.
- */
-enum LanguagesEnum {
     /**
-     * Html language
+     * Summary Function that does all the steps to minify all the css code.
+     * @param cssContent css array to be minified.
+     * @return string with all the code minified.
      */
-    HTML = '.html',
+    public minifyCssScssLessSass(cssContent: string[]): string {
+        const removeComments: string[] = this.removeComments(cssContent);
+        const hexMinifiedCss: string[] = this.HexMinify(removeComments);
+        const minifierCss: CssMinifier = new CssMinifier(hexMinifiedCss);
+        return minifierCss.getCssMinified();
+    }
+
     /**
-     * Css language
+     * Summary Function that does all the steps to minify all the json code.
+     * @param jsonContent json array to be minified.
+     * @return  string with all the code minified.
      */
-    CSS = '.css',
+    public minifyJsonJsonc(jsonContent: string[]): string {
+        const contentWithHexMinified: string[] = this.HexMinify(jsonContent);
+        const removeComments: string[] = this.removeComments(contentWithHexMinified);
+        const minifierJson: JsonMinifier = new JsonMinifier(removeComments);
+        return minifierJson.getJSONMinified();
+    }
+
     /**
-     * Json language
+     * Summary Function that does all the steps to minify all the html code.
+     * @param  htmlContent html array to be minified.
+     * @return string with all the code minified.
      */
-    JSON = '.json',
+    public minifyHtml(htmlContent: string[]): string {
+        const minifierHtml: HtmlMinifier = new HtmlMinifier(htmlContent);
+        minifierHtml.removeMultipleLineComments();
+        return minifierHtml.getHtmlMinified();
+    }
+
     /**
-     * Jsonc language
+     * Summary minifies hexadecimal code if enabled.
+     *
+     * Description receives an array with all the content,
+     * and minifies it's hexadecimal, rgb and rgba values;
+     * then return the new array;
+     * If it is enabled it will initialize the HexMinifier
+     * class and it will make all the processes and return
+     * the new array of values OR it will simply return the
+     * received value and do nothing.
+     *
+     * @param content Array with all the lines to be hexMinified.
+     * @return content with the colors minified.
      */
-    JSONC = '.jsonc',
+    public HexMinify(content: string[]): string[] {
+        let returnValue: string[];
+
+        if (this.willMinifyHex) {
+            const hexMinifier: HexMinifier = new HexMinifier(content);
+            // Minifier methods
+            hexMinifier.shortHexMain();
+            hexMinifier.shortRGBMain();
+            hexMinifier.shortRGBAMain();
+            returnValue = hexMinifier.getHexMinified();
+        } else {
+            returnValue = content;
+        }
+        return returnValue;
+    }
+
+
+    /**
+     * Summary it removes the comments from a class. (not for HTML).
+     *
+     * Description removeComments receives an array with the content
+     * and removes single line and multiple line comments (//)(/* * /)
+     * by calling the class removeComments and calling the method
+     * removeCommentsMain, Then gets the result with getLineRemoved.
+     *
+     * @param content All the content to remove the comments
+     * @return content without the comments.
+     */
+    public removeComments(content: string[]): string[] {
+        const removeComments = new CommentRemover(content);
+        removeComments.removeCommentsMain();
+        return removeComments.getCommentsRemoved();
+    }
 }
